requirements.txt
cffi==1.5.2
greenlet==0.4.9
MySQL-python==1.2.5
numpy==1.10.0
python-dateutil==2.7.5
pytz==2018.7
readline==6.2.4.1
six==1.11.0
